# VETERINARIAPROD/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  # ── AUTH ────────────────────────────────────────────────────────────────────
paths:
...
  /lab/orders:
    get:
      tags: [Lab]
      summary: Listar órdenes de laboratorio
      parameters:
        - { in: query, name: patientId, schema: { type: integer } }
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Lista de órdenes }
    post:
      tags: [Lab]
      summary: Crear orden de laboratorio
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [patientId, tests]
              properties:
                patientId:       { type: integer }
                medicalRecordId: { type: integer }
                tests:
                  type: array
                  description: "IDs de lab_tests o items con testId"
                  items:
                    oneOf:
                      - type: integer
                      - type: object
                        required: [testId]
                        properties:
                          testId: { type: integer }
                priority:        { type: string, enum: [routine,urgent,stat], default: routine }
      responses:
        '201': { description: Orden creada con número auto-generado }
...
  /lab/orders/{id}:
    get:
      tags: [Lab]
      summary: Obtener detalle de una orden de laboratorio
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Orden con items y resultados }
        '404': { description: Orden no encontrada }
...
  /vaccinations/vaccines:
    get:
      tags: [Vaccination]
      summary: Catalogo de vacunas
      parameters:
        - in: query
          name: speciesId
          schema: { type: integer }
      responses:
        '200': { description: Lista de vacunas }
...
  /invoices:
    get:
      tags: [Billing]
      summary: Listar facturas
      parameters:
        - in: query
          name: page
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: limit
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: status
          schema: { type: string, enum: [draft,pending,partial,paid,overdue,cancelled] }
        - in: query
          name: clientId
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Lista paginada de facturas }
    post:
      tags: [Billing]
      summary: Crear factura
      description: Usa sp_generate_invoice. Numeración auto-correlativa por sucursal.
      responses:
        '201': { description: Factura creada }
...
  /billing/consolidated/summary:
    get:
      tags: [Billing]
      summary: Resumen financiero consolidado multi-sucursal
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Resumen consolidado por período }
...
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Marcar notificación como leída
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Notificación leída }
...
  /portal/pets/{id}/vaccinations:
    get:
      tags: [Portal]
      summary: Vacunas de la mascota
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Vacunas }
...
  /portal/notifications/{id}/read:
    patch:
      tags: [Portal]
      summary: Marcar notificación de portal como leída
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Notificación leída }
...